Handle visual in a more compatible way
Marks `<` and `>` are only populated on exiting visual mode, which
kinda works with `lua` cmdline call, as it forcibly exists visual mode.

But keybindings don't do that, and as a result the plugin never
accesses the visual selection that user is looking at.

This commit relegates marks usage to a cmdline mode, uses getpos
instead, and implements a `visual` option to render visual portion
of the buffer.